// dbschema/interfaces.ts
export namespace std {
  export interface BaseObject {
    "id": string;
  }
  export interface $Object extends BaseObject {}
  export interface FreeObject extends BaseObject {}
  export type JsonEmpty = "ReturnEmpty" | "ReturnTarget" | "Error" | "UseNull" | "DeleteKey";
}
export namespace auth {
  export interface Branch extends std.$Object {
    "company"?: Company | null;
    "tenant": Tenant;
    "users": User[];
    "active": boolean;
    "created_at": Date;
    "description"?: string | null;
    "name": string;
    "slug": string;
  }
  export interface Company extends std.$Object {
    "tenant": Tenant;
    "branches": Branch[];
    "description"?: string | null;
    "created_at": Date;
    "name": string;
    "active": boolean;
    "slug": string;
  }
  export interface EmployeeRole extends std.$Object {
    "permissions": Permission[];
    "users": User[];
    "tenant": Tenant;
    "description"?: string | null;
    "active": boolean;
    "name": string;
  }
  export interface Permission extends std.$Object {
    "employee_roles": EmployeeRole[];
    "description"?: string | null;
    "active": boolean;
    "name": string;
    "slug": string;
  }
  export interface Tenant extends std.$Object {
    "tenant_metadata"?: TenantMetadata | null;
    "active"?: boolean | null;
    "description"?: string | null;
    "name": string;
    "created_at": Date;
  }
  export interface TenantMetadata extends std.$Object {
    "tenant_favicon"?: string | null;
    "tenant_logo"?: string | null;
    "tenant_main_color"?: string | null;
    "tenant_secondary_color"?: string | null;
  }
  export interface User extends std.$Object {
    "tenant": Tenant;
    "employee_role"?: EmployeeRole | null;
    "branch": Branch;
    "branches": Branch[];
    "name"?: string | null;
    "pin"?: string | null;
    "password": string;
    "username": string;
    "active": boolean;
    "created_at": Date;
    "salt": string;
    "user_role": UserRole;
    "slug"?: string | null;
  }
  export type UserRole = "Admin" | "Employee";
}